requirements.txt
Pillow>=10.1.0
pillow-heif>=0.22.0